Test for and use the GotoBLAS2 in I3_PORTS.
Now that GotoBLAS2 is in ports, and our port of SuiteSparse requires it, prefer
that one to others. If it's not found, do a normal search. Hopefully, the
user has a good reason for not using GotoBLAS2 and has overridden all this in
I3_SITE_CMAKE anyway.

Additionally, only look for BLAS/LAPACK if we're building the icerec meta
project, since wavedeform is the only project that uses it (via SuiteSparse).


git-svn-id: http://code.icecube.wisc.edu/icetray/projects/cmake/trunk@1004 579d543e-85e3-4578-a5e8-1e6a993b7602