relax the forcing of cmake variables
The gratuitous use of set(... FORCE) causes changes deep in the dependency
tree. This triggers a "touch" of files that the Makefile generator uses,
which in turn causes make to think that projects are out of date, forcing
a rebuild when it's not necessary.

I was able to reproduce this consistantly in two rare situations, and others
have seen this behavior as well. This patch at least fixes the two situations
I observed.


git-svn-id: http://code.icecube.wisc.edu/icetray/projects/cmake/trunk@2193 579d543e-85e3-4578-a5e8-1e6a993b7602