fix the default warning suppression flag tests
GCC doesn't flag unknown warning suppression flags at low warning levels.
you can litterally use "-Wno-poop" and GCC will ignore it. this obviously
doesn't work for testing single flags. So, we change the test logic to look
for the warning ("-Wpoop") if that exists, then we enable the suppression,
because we assume that suppression doesn't exist without it's corresponding
warning.


git-svn-id: http://code.icecube.wisc.edu/icetray/projects/cmake/trunk@2415 579d543e-85e3-4578-a5e8-1e6a993b7602