clean out cruft from the macro `i3_test-script()`
this cruft supported the old test framework and snowblower, and involved a
lot of writing and scraping of files. now, with ctest, this is just unused
code and bloats the macro making it harder to read. the macro 
`i3_test_executable()` contains similar cruft, but is a more complicated
macro and needs to be pruned more carefully.


git-svn-id: http://code.icecube.wisc.edu/icetray/projects/cmake/trunk@2622 579d543e-85e3-4578-a5e8-1e6a993b7602