clean out cruft from the macro `i3_test_executable()`
this cruft supported the old test framework and snowblower, and involved a
lot of writing and scraping of files. now, with ctest, this is just unused
code and bloats the macro making it harder to read.

this cruft also allowed using the individual tests in each binary test
driver; something we haven't done for quite some time. since moving to ctest
we just run all tests in the binary test driver at once.


git-svn-id: http://code.icecube.wisc.edu/icetray/projects/cmake/trunk@2624 579d543e-85e3-4578-a5e8-1e6a993b7602